# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZICK — A spiritual-wellness app backend providing auth, content, schedules, and subscriptions.",
        "title": "ZICK API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ZICK API Support",
            "email": "support@zick.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/library": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to create a new library item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Create a library item",
                "parameters": [
                    {
                        "description": "Library payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.CreateLibraryReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.LibraryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/library-categories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to create a new category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library Categories"
                ],
                "summary": "Create a library category",
                "parameters": [
                    {
                        "description": "Category payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.CreateCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/library-categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to update a category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library Categories"
                ],
                "summary": "Update a library category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.UpdateCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to delete a category.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library Categories"
                ],
                "summary": "Delete a library category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/admin/library/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to update an existing library item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Update a library item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Library Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Library payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.UpdateLibraryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.LibraryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to delete an existing library item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Delete a library item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Library Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/library": {
            "get": {
                "description": "Retrieves a paginated list of library items. Supports search, sort, category filtering, and pagination via query params.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Get all library items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by title, category, or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort column; prefix with - for DESC (e.g. -created_at)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.PaginatedLibraryResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/library/{id}": {
            "get": {
                "description": "Retrieves details of a specific library item along with related items.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "Get library item details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Library Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_library_dto.LibraryDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gotickets_internal_domain_library_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "gotickets_internal_domain_library_dto.CreateCategoryReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "gotickets_internal_domain_library_dto.CreateLibraryReq": {
            "type": "object",
            "required": [
                "category",
                "content_text",
                "short_description",
                "thumbnail_url",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "maxLength": 100
                },
                "content_text": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "gotickets_internal_domain_library_dto.LibraryDetailsResponse": {
            "type": "object",
            "properties": {
                "library_item": {
                    "$ref": "#/definitions/gotickets_internal_domain_library_dto.LibraryResponse"
                },
                "related": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gotickets_internal_domain_library_dto.LibraryResponse"
                    }
                }
            }
        },
        "gotickets_internal_domain_library_dto.LibraryResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "content_text": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "gotickets_internal_domain_library_dto.PaginatedLibraryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gotickets_internal_domain_library_dto.LibraryResponse"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/gotickets_internal_querybuilder.Meta"
                }
            }
        },
        "gotickets_internal_domain_library_dto.UpdateCategoryReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "gotickets_internal_domain_library_dto.UpdateLibraryReq": {
            "type": "object",
            "required": [
                "category",
                "content_text",
                "short_description",
                "thumbnail_url",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "maxLength": 100
                },
                "content_text": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "gotickets_internal_httpresponse.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "gotickets_internal_querybuilder.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Limit is the number of items per page.",
                    "type": "integer"
                },
                "page": {
                    "description": "Page is the current page number (1-indexed).",
                    "type": "integer"
                },
                "total": {
                    "description": "Total is the total number of matching records.",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "TotalPages is the total number of pages.",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}